Split gardens repository modules
