+++ packages/storage/src/index.ts
@@ -60,6 +60,9 @@ export * from './repositories/outletOffersRepo';
 export * from './repositories/payoutsRepo';
 export * from './repositories/pickupLocationsRepo';
 export * from './repositories/plantRelationshipsRepo';
+export * from './repositories/raisedBedDiaryRepo';
+export * from './repositories/raisedBedFieldsRepo';
+export * from './repositories/raisedBedsRepo';
 export * from './repositories/referralsRepo';
 export * from './repositories/refreshTokensRepo';
 export * from './repositories/seasonalOffersRepo';
